Use mpmc instead of mpsc if available
The mpmc (Multiple Producer Multiple Consumer) module
of the rust std lib, is a better fit for a job pool.

mpsc has a single consumer and multiple producers. A job pool is the
oposite. A single producer (the pool) and multiple consumers (the
workers).

Check the rust compiler version and enable the mpmc feature gate if available.

This avoids the need to wrap the Receiver<T> inside an Arc + Mutex.