Remove 'builders' dependency
The `PoolConfig` struct is very simple. Using a derive macro to generate
a builder that can be manually implemented in 40 lines of code is overkill.

This debloats the crate, making it dependency free, expect for the
bindings feature, which still depends on cbindgen at build time.

Also, now the builder functions are `const`.